+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/cdi/CustomPassivatingNormalScope.java
@@ -1,0 +1,15 @@
+package io.openliberty.sample.jakarta.cdi;
+
+import jakarta.enterprise.context.NormalScope;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import static java.lang.annotation.ElementType.*;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@NormalScope(passivating = true)
+@Retention(RUNTIME)
+@Target({TYPE, METHOD, FIELD})
+public @interface CustomPassivatingNormalScope {
+}
