+++ src/test/resources/projects/maven/jakarta-sample/src/main/java/io/openliberty/sample/jakarta/cdi/CustomPassivatingScopeWithoutSerializable.java
@@ -1,0 +1,5 @@
+package io.openliberty.sample.jakarta.cdi;
+
+@CustomPassivatingNormalScope
+public class CustomPassivatingScopeWithoutSerializable {
+}
